Step 2 (consolidation): CombinationCost + naming/structure alignment
Byte-identical consolidation after the five theory conversions:

- New CombinationCost holder for costs whose meaning spans strategies, with two
  documented admission mechanisms: (1) conflict-dispatched rule sets
  (ORDERED_REWRITING for apply_equations - the demodulation cost, guarded by
  term/monomial-ordering decrease at all four call sites; CNF_CONVERSION for
  apply_equations_andOr, shared deliberately by FOL's conjNormalForm), and
  (2) cost-sum couplings (APPLY_SELECT_EQ_EFFECTIVE = tuned sum of the dual-
  tagged applyEq taclet; HeapSelectCost.APPLY_SELECT_EQ is now derived from it,
  turning the old "about -5700" comment into checked arithmetic).
  FOLCost.APPLY_EQUATIONS/CNF_ORDERING and LinearEquationCost.APPLY_EQUATIONS/
  APPLY_EQ_AND_OR are removed in favour of the shared constants.
- JavaCardDL holders grouped per theory into JavaCardDLCosts.java
  (JavaCardDLCost + HeapSelectCost, Integer-style one file per theory).
- Below-scope penalty names aligned: StringCost.BELOW_MODALITY (matching
  DivModCost.BELOW_MODALITY), SymExCost.PROGRAM_STEP_BELOW_QUANTIFIER.

A sweep over all taclets and rule-set bindings verified completeness: exactly
the three conflict-dispatched rule sets exist (order_terms, apply_equations,
apply_equations_andOr, all FOL/Integer), and the only tuned cost-sum coupling
is the applyEq taclet covered above.

Created with AI tooling support
